Combine season stat classes to one class and change that in stat_tracker respectively
