+++ test/season_stat_test.rb
@@ -11,9 +11,11 @@ class SeasonStatTest < Minitest::Test
   def setup
     team_file_path = './data/teams.csv'
     game_file_path = './test/fixtures/truncated_games.csv'
+    game_team_file_path = './test/fixtures/truncated_game_teams.csv'
     @game_collection = GameCollection.new(game_file_path)
     @team_collection = TeamCollection.new(team_file_path)
-    @season_stat = SeasonStat.new(@game_collection, @team_collection)
+    @game_team_collection = GameTeamCollection.new(game_team_file_path)
+    @season_stat = SeasonStat.new(@game_collection, @team_collection, @game_team_collection)
 
     @team_info = {
 
@@ -78,7 +80,7 @@ class SeasonStatTest < Minitest::Test
 
     assert_equal 257, @season_stat.count_of_season_games("20122013")
   end
-
+#
   def test_it_can_get_average_goals_by_season
     @season_stat.get_all_seasons
     @season_stat.season_games_by_all_seasons
@@ -170,5 +172,160 @@ class SeasonStatTest < Minitest::Test
 
     @season_stat.stubs(:get_team_data).returns(@team_info)
     assert_equal "Apples", @season_stat.biggest_surprise(@season)
+  end
+#
+  def test_it_can_get_coaches_by_season
+    @season_stat.get_all_seasons
+    @season_stat.season_game_teams_by_all_seasons
+
+    assert_instance_of Array, @season_stat.coaches_by_season("20122013")
+    assert_equal "John Tortorella", @season_stat.coaches_by_season("20122013")[0]
+    assert_equal "Claude Julien", @season_stat.coaches_by_season("20122013")[1]
+    assert_equal "Dan Bylsma", @season_stat.coaches_by_season("20122013")[2]
+    assert_equal "Mike Babcock", @season_stat.coaches_by_season("20122013")[3]
+    assert_equal "Joel Quenneville", @season_stat.coaches_by_season("20122013")[4]
+    assert_equal "Paul MacLean", @season_stat.coaches_by_season("20122013")[5]
+    assert_equal "Michel Therrien", @season_stat.coaches_by_season("20122013")[6]
+    assert_equal "Mike Yeo", @season_stat.coaches_by_season("20122013")[7]
+    assert_equal "Darryl Sutter", @season_stat.coaches_by_season("20122013")[8]
+    assert_equal "Ken Hitchcock", @season_stat.coaches_by_season("20122013")[9]
+    assert_equal "Bruce Boudreau", @season_stat.coaches_by_season("20122013")[10]
+    assert_equal "Jack Capuano", @season_stat.coaches_by_season("20122013")[11]
+    assert_equal "Adam Oates", @season_stat.coaches_by_season("20122013")[12]
+    assert_equal "Todd Richards", @season_stat.coaches_by_season("20122013")[13]
+    assert_equal 14, @season_stat.coaches_by_season("20122013").length
+    assert_nil @season_stat.coaches_by_season("20122013")[29]
+  end
+
+  def test_it_can_get_coach_wins_by_season
+    @season_stat.get_all_seasons
+    @season_stat.season_game_teams_by_all_seasons
+
+    assert_equal 2, @season_stat.get_coach_wins_by_season("John Tortorella", "20122013")
+    assert_equal 9, @season_stat.get_coach_wins_by_season("Claude Julien", "20122013")
+    assert_equal 4, @season_stat.get_coach_wins_by_season("Dan Bylsma", "20122013")
+    assert_equal 7, @season_stat.get_coach_wins_by_season("Mike Babcock", "20122013")
+    assert_equal 9, @season_stat.get_coach_wins_by_season("Joel Quenneville", "20122013")
+    assert_equal 3, @season_stat.get_coach_wins_by_season("Paul MacLean", "20122013")
+    assert_equal 1, @season_stat.get_coach_wins_by_season("Michel Therrien", "20122013")
+    assert_equal 1, @season_stat.get_coach_wins_by_season("Mike Yeo", "20122013")
+    assert_equal 5, @season_stat.get_coach_wins_by_season("Darryl Sutter", "20122013")
+    assert_equal 3, @season_stat.get_coach_wins_by_season("Ken Hitchcock", "20122013")
+    assert_equal 4, @season_stat.get_coach_wins_by_season("Bruce Boudreau", "20122013")
+    assert_equal 2, @season_stat.get_coach_wins_by_season("Jack Capuano", "20122013")
+    assert_equal 5, @season_stat.get_coach_wins_by_season("Adam Oates", "20122013")
+    assert_equal 0, @season_stat.get_coach_wins_by_season("Todd Richards", "20122013")
+  end
+
+  def test_it_can_get_all_coach_games_by_season
+    @season_stat.get_all_seasons
+    @season_stat.season_game_teams_by_all_seasons
+
+    assert_equal 12, @season_stat.get_total_coach_games_by_season("John Tortorella", "20122013")
+  end
+
+  def test_it_can_get_coach_win_percentage_by_season
+    @season_stat.get_all_seasons
+    @season_stat.season_game_teams_by_all_seasons
+
+    assert_equal 16.67, @season_stat.coach_win_percentage_by_season("John Tortorella", "20122013")
+  end
+
+  def test_it_can_create_coaches_hash
+    @season_stat.get_all_seasons
+    @season_stat.season_game_teams_by_all_seasons
+
+    assert_instance_of Hash, @season_stat.create_coach_win_data_by_season("20122013")
+    assert_equal 16.67, @season_stat.create_coach_win_data_by_season("20122013")["John Tortorella"]
+  end
+
+  def test_it_can_find_winningest_coach
+    @season_stat.get_all_seasons
+    @season_stat.season_game_teams_by_all_seasons
+
+    assert_equal "Claude Julien", @season_stat.winningest_coach("20122013")
+  end
+
+  def test_it_can_find_worst_coach
+    @season_stat.get_all_seasons
+    @season_stat.season_game_teams_by_all_seasons
+
+    assert_equal "Todd Richards", @season_stat.worst_coach("20122013")
+  end
+
+  def test_it_can_get_team_ids_by_season
+    @season_stat.get_all_seasons
+    @season_stat.season_game_teams_by_all_seasons
+
+    assert_instance_of Array, @season_stat.get_team_ids_by_season("20122013")
+    assert_equal 14, @season_stat.get_team_ids_by_season("20122013").length
+    assert_equal 3, @season_stat.get_team_ids_by_season("20122013")[0]
+  end
+
+  def test_it_can_get_team_name_by_team_id
+    @season_stat.get_all_seasons
+    @season_stat.season_game_teams_by_all_seasons
+
+    assert_equal "Houston Dynamo", @season_stat.get_team_name(3)
+  end
+
+  def test_it_can_get_team_tackles_by_season
+    @season_stat.get_all_seasons
+    @season_stat.season_game_teams_by_all_seasons
+    assert_instance_of Integer, @season_stat.get_tackles_by_team_season("3", "20122013")
+    assert_equal 466, @season_stat.get_tackles_by_team_season("3", "20122013")
+  end
+
+  def test_it_can_get_team_goals_by_season
+    @season_stat.get_all_seasons
+    @season_stat.season_game_teams_by_all_seasons
+    assert_equal 18, @season_stat.get_goals_by_team_season("3", "20122013")
+  end
+
+  def test_it_can_get_team_shots_by_season
+    @season_stat.get_all_seasons
+    @season_stat.season_game_teams_by_all_seasons
+    assert_equal 87, @season_stat.get_shots_by_team_season("3", "20122013")
+  end
+
+  def test_it_can_get_goals_to_shots_ratio
+    @season_stat.get_all_seasons
+    @season_stat.season_game_teams_by_all_seasons
+    assert_equal 4.833, @season_stat.team_shots_to_goal_ratio_by_season("3", "20122013")
+  end
+
+  def test_it_can_create_team_data_by_season
+    @season_stat.get_all_seasons
+    @season_stat.season_game_teams_by_all_seasons
+    assert_instance_of Hash, @season_stat.create_team_data_by_season("20122013")
+    assert_equal 4.833, @season_stat.create_team_data_by_season("20122013")["3"][:goal_ratio]
+    assert_equal "Houston Dynamo", @season_stat.create_team_data_by_season("20122013")["3"][:team_name]
+    assert_equal 466, @season_stat.create_team_data_by_season("20122013")["3"][:tackles]
+    assert_nil @season_stat.create_team_data_by_season("20122013")["100"]
+    assert_equal 14, @season_stat.create_team_data_by_season("20122013").keys.length
+  end
+
+  def test_it_can_find_most_accurate_team_name
+    @season_stat.get_all_seasons
+    @season_stat.season_game_teams_by_all_seasons
+    assert_equal "New York City FC", @season_stat.most_accurate_team("20122013")
+  end
+
+  def test_it_can_find_least_accurate_team_name
+    @season_stat.get_all_seasons
+    @season_stat.season_game_teams_by_all_seasons
+    assert_equal "Houston Dynamo", @season_stat.least_accurate_team("20122013")
+  end
+
+  def test_it_can_get_team_with_most_tackles
+    @season_stat.get_all_seasons
+    @season_stat.season_game_teams_by_all_seasons
+    assert_equal "Houston Dynamo", @season_stat.most_tackles("20122013")
+  end
+
+  def test_it_can_find_team_with_least_tackles
+    @season_stat.get_all_seasons
+    @season_stat.season_game_teams_by_all_seasons
+    assert_equal "Orlando Pride", @season_stat.fewest_tackles("20122013")
   end
 end
